requirements.txt
beautifulsoup4
lxml
requests
schedule
pymongo
python-dotenv


